Refactoring  componets
