--- src/components/atoms/Button/button.scss
@@ -1,80 +1,0 @@
-$main-color: #007cc7;
-$withe-color: #fff;
-$grey-light-color: #ededed;
-$grey-darc-color: #919191;
-$base-font-family: sans-serif;
-
-.button {
-  -webkit-appearance: none;
-  border: 0;
-  outline: 0;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  font-family: $base-font-family;
-  font-size: 0.9em;
-  font-weight: 600;
-  text-transform: uppercase;
-  border-radius: 3px;
-  cursor: pointer;
-  position: relative;
-
-  &:hover {
-    color: $main-color;
-    background: $withe-color;
-  }
-     &:disabled {
-    color: $grey-darc-color;
-    background: $withe-color;
-    border-color: $grey-light-color;
-    cursor: default;
-  }
-}
-
-/* Form Icon Button */
-.circle {
-  height: 2.7rem;
-  width: 2.7rem;
-  border-radius: 50%;
-  display: inline-block;
-}
-
-/* Size Button */
-.small {
-  padding: 0.5em 0.5em;
-}
-
-.medium {
-  padding: 0.8em 1.1em;
-}
-
-.large {
-  padding: 1.7em 2.4em;
-}
-
-/* Theme */
-.default {
-  color: $withe-color;
-  background: $main-color;
-  border: 1px solid $main-color;
-}
-
-.rounded {
-  color: $main-color;
-  background: $withe-color;
-  border: 1px solid $main-color;
-  transition: color 0.3s ease, background 0.3s ease;
-
-  &:hover {
-    color: $withe-color;
-    background: $main-color;
-  }
-
-  &:disabled {
-    color: $grey-darc-color;
-    background: $withe-color;
-    border-color: $grey-light-color;
-    cursor: default;
-  }
-}
