+++ src/components/atoms/IconButton/IconButton.jsx
@@ -2,64 +2,54 @@ import React from 'react'
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
 
-import styles from './button.scss'
+import styles from './iconButton.module.scss'
 
-export const ButtonType = {
+export const IconType = {
   BUTTON: 'button',
   RESET: 'reset',
   SUBMIT: 'submit',
 }
 
-export const ButtonTheme = {
+export const IconTheme = {
   DEFAULT: 'default',
   ROUNDED: 'rounded',
 }
 
-export const ButtonForm = {
-  CIRCLE: 'circle',
-}
-
-export const ButtonSize = {
+export const IconSize = {
   SMALL: 'small',
   MEDIUM: 'medium',
   LARGE: 'large',
 }
 
-const Button = ({ icon, type, onClick, children, theme, form, size, className, disabled }) => {
+export const IconButton = ({ icon, type, onClick, theme, size, className, disabled }) => {
   const iconChildren = <i className={icon} />
 
-  const classProps = classNames(styles.button, styles[theme], styles[size], styles[form], className)
+  const classProps = classNames(styles.button, styles[theme], styles[size], className)
 
   return (
     <button type={type} onClick={onClick} disabled={disabled} className={classProps}>
       {icon && iconChildren}
-      {children}
     </button>
   )
 }
 
-Button.propTypes = {
+IconButton.propTypes = {
   icon: PropTypes.string,
   type: PropTypes.string,
-  form: PropTypes.string,
   size: PropTypes.string,
   theme: PropTypes.string,
   disabled: PropTypes.bool,
-  children: PropTypes.string,
   className: PropTypes.string,
   onClick: PropTypes.func.isRequired,
 }
 
-Button.defaultProps = {
+IconButton.defaultProps = {
   icon: '',
-  form: '',
-  size: '',
-  children: '',
   className: '',
   disabled: false,
   onClick: () => {},
-  type: ButtonType.BUTTON,
-  theme: ButtonTheme.DEFAULT,
+  size: IconSize.MEDIUM,
+  type: IconType.BUTTON,
+  theme: IconTheme.DEFAULT,
 }
 
-export { Button }
