--- src/components/atoms/IconButton/iconButton.scss
@@ -1,72 +1,0 @@
-$main-color: #007cc7;
-$withe-color: #fff;
-$grey-light-color: #ededed;
-$grey-darc-color: #919191;
-$base-font-family: sans-serif;
-
-.button {
-  -webkit-appearance: none;
-  border: 0;
-  outline: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  position: relative;
-  border-radius: 50%;
-
-  &:hover {
-    color: $main-color;
-    background: $withe-color;
-  }
-     &:disabled {
-    color: $grey-darc-color;
-    background: $withe-color;
-    border-color: $grey-light-color;
-    cursor: default;
-  }
-}
-
-/* Size */
-.small {
-  height: 2.1rem;
-  width: 2.1rem;
-  
-}
-
-.medium {
-  height: 2.7rem;
-  width: 2.7rem;
-}
-
-.large {
-  height: 4.3rem;
-  width: 4.3rem;
-}
-
-/* Theme */
-.default {
-  color: $withe-color;
-  background: $main-color;
-  border: 1px solid $main-color;
-}
-
-.rounded {
-  color: $main-color;
-  background: $withe-color;
-  border: 1px solid $main-color;
-  transition: color 0.3s ease, background 0.3s ease;
-
-  &:hover {
-    color: $withe-color;
-    background: $main-color;
-  }
-
-  &:disabled {
-    color: $grey-darc-color;
-    background: $withe-color;
-    border-color: $grey-light-color;
-    cursor: default;
-  }
-}
-
