+++ src/components/atoms/Input/index.js
@@ -1,0 +1,2 @@
+export { Input } from "./Input";
+
